+++ tests/test_ir.py
@@ -997,3 +997,23 @@ def test_replace_by_if():
 
     assert set(u.operation for u in a.uses) == {c}
     assert set(u.operation for u in d.uses) == {b}
+
+
+def test_find_ancestor_op_in_block_has_ancestor():
+    op1 = test.TestOp()
+    op2 = test.TestOp(regions=[Region(Block([op1]))])
+
+    blk_top = Block([op2])
+    test.TestOp(regions=[Region(blk_top)])
+
+    assert blk_top.find_ancestor_op_in_block(op1) is op2
+
+
+def test_find_ancestor_op_in_block_has_no_ancestor():
+    op1 = test.TestOp()
+    op2 = test.TestOp(regions=[Region(Block([op1]))])
+
+    blk_top = Block([op2])
+    op3 = test.TestOp(regions=[Region(blk_top)])
+
+    assert blk_top.find_ancestor_op_in_block(op3) is None
